delete methods and activities which are not needed for TicketCheck
