+++ app/build.gradle
@@ -31,7 +31,7 @@ android {
     }
     defaultConfig {
         applicationId "de.tum.in.tca.ticketcheck"
-        minSdkVersion 24
+        minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
         versionName "1.0"
