+++ app/src/main/java/de/tum/in/tca/ticketcheck/api/TUMCabeAPIService.java
@@ -35,28 +35,10 @@ public interface TUMCabeAPIService {
     @GET(API_EVENTS + "list/search/{searchTerm}")
     Call<List<Event>> searchEvents(@Path("searchTerm") String searchTerm);
 
-    // Getting Ticket information
-    @POST(API_EVENTS + API_TICKET + "my")
-    Call<List<Ticket>> getTickets(@Body ChatVerification chatVerification);
-
     @GET(API_EVENTS + API_TICKET + "type/{eventID}")
     Call<List<TicketType>> getTicketTypes(@Path("eventID") int eventID);
 
-    // Ticket reservation
-    @POST(API_EVENTS + API_TICKET + "reserve")
-    Call<TicketReservationResponse> reserveTicket(@Body ChatVerification chatVerification);
-
-    @POST(API_EVENTS + API_TICKET + "reserve/cancel")
-    Call<TicketSuccessResponse> cancelTicketReservation(@Body ChatVerification chatVerification);
-
-    // Ticket purchase
-    @POST(API_EVENTS + API_TICKET + "payment/stripe/purchase")
-    Call<Ticket> purchaseTicketStripe(@Body ChatVerification chatVerification);
-
     @POST(API_EVENTS + "ticket/validate")
     Call<TicketValidityResponse> getNameForTicket(@Body TicketValidityRequest request);
-
-    @POST(API_EVENTS + API_TICKET + "payment/stripe/ephemeralkey")
-    Call<HashMap<String, Object>> retrieveEphemeralKey(@Body ChatVerification chatVerification);
 
 }
