+++ app/src/main/java/de/tum/in/tca/ticketcheck/component/ticket/EventsController.java
@@ -23,8 +23,6 @@ public class EventsController {
     private final Context context;
 
     private final EventDao eventDao;
-    private final TicketDao ticketDao;
-    private final TicketTypeDao ticketTypeDao;
 
     /**
      * Constructor, open/create database, create table if necessary
@@ -34,8 +32,6 @@ public class EventsController {
     public EventsController(Context context) {
         this.context = context;
         eventDao = TcaDb.getInstance(context).eventDao();
-        ticketDao = TcaDb.getInstance(context).ticketDao();
-        ticketTypeDao = TcaDb.getInstance(context).ticketTypeDao();
     }
 
 
@@ -52,37 +48,6 @@ public class EventsController {
         } catch (IOException e) {
             Utils.log(e);
         }
-
-        // Load all tickets
-        try {
-            if(Utils.getSetting(context, Const.CHAT_MEMBER, ChatMember.class) != null) {
-                api.getTickets(context, new Callback<List<Ticket>>() {
-                    @Override
-                    public void onResponse(Call<List<Ticket>> call, Response<List<Ticket>> response) {
-                        List<Ticket> list = response.body();
-                        if (list == null) list = new ArrayList<Ticket>();
-                        ticketDao.insert(list);
-                    }
-
-                    @Override
-                    public void onFailure(Call<List<Ticket>> call, Throwable t) {
-                        //TODO: inform user about failure
-                        t.printStackTrace();
-                    }
-                });
-            }
-        } catch (IOException e) {
-            Utils.log(e);
-        }
-
-        // Load all ticket types
-        try {
-            // TODO: replace by real event ids! -> loop over all found ids
-            List ticketTypes = api.getTicketTypes(1);
-            //ticketTypeDao.insert(ticketTypes);
-        } catch (IOException e) {
-            Utils.log(e);
-        }
     }
 
     // Event methods
@@ -91,58 +56,8 @@ public class EventsController {
         return eventDao.getAll();
     }
 
-    public List<Event> getBookedEvents() {
-        // Return all events for which a ticket exists
-        List<Ticket> tickets = ticketDao.getAll();
-        List<Event> bookedEvents = new ArrayList<>();
-        for (Ticket ticket : tickets){
-            bookedEvents.add(getEventById(ticket.getEventId()));
-        }
-        return bookedEvents;
-    }
-
-    public boolean isEventBooked(Event event) {
-        for (Event bookedEvent : getBookedEvents()) {
-            if (bookedEvent.getId() == event.getId()) {
-                return true;
-            }
-        }
-        return false;
-    }
-
     public Event getEventById(int id) {
         return eventDao.getEventById(id);
-    }
-
-    // Ticket methods
-
-    public Ticket getTicketByEventId(int eventId) {
-        return ticketDao.getByEventId(eventId);
-    }
-
-    public TicketType getTicketTypeById(int id) {
-        return ticketTypeDao.getById(id);
-    }
-
-    /**
-     * This is not a database access but a API call
-     * Thus, it needs to be called in a thread
-     * @param eventId
-     * @return
-     */
-    public List<TicketType> getTicketTypesByEventId(int eventId) {
-        List<TicketType> ticketTypes = null;
-        try {
-            TUMCabeClient api = TUMCabeClient.getInstance(context);
-            ticketTypes = api.getTicketTypes(eventId);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return ticketTypes;
-    }
-
-    public void addTickets(List<Ticket> tickets) {
-        ticketDao.insert(tickets);
     }
 }
