--- app/src/main/res/layout/event_fragment.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:background="@color/white"
-    android:gravity="center"
-    android:orientation="vertical">
-
-    <android.support.v4.widget.SwipeRefreshLayout
-        android:id="@+id/event_refresh"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <!--  Recycler View  -->
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/activity_events_list_view"
-            android:layout_width="fill_parent"
-            android:layout_height="fill_parent"
-            android:scrollbars="none" />
-    </android.support.v4.widget.SwipeRefreshLayout>
-</LinearLayout>
